dyn_exp_cover/requirements.txt
requests~=2.26.0
Pillow~=8.4.0
vk-api~=11.9.7